+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.0.5
 
-- Fixed a bug on location-based statically add of places via
+- Fixed a bug on location-based statically add of places via script
 
 # 2.0.4